refactor: merge poly and lookup under baseconfig
