+++ benches/accum_conv.rs
@@ -1,5 +1,5 @@
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion, Throughput};
-use ezkl_lib::circuit::base::*;
+use ezkl_lib::circuit::*;
 use ezkl_lib::commands::TranscriptType;
 use ezkl_lib::execute::create_proof_circuit_kzg;
 use ezkl_lib::pfsys::create_keys;
@@ -67,7 +67,8 @@ impl Circuit<Fr> for MyCircuit {
                         Op::Conv {
                             padding: (0, 0),
                             stride: (1, 1),
-                        },
+                        }
+                        .into(),
                     )
                     .unwrap();
                 Ok(())
