+++ benches/accum_sumpool.rs
@@ -1,5 +1,5 @@
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion, Throughput};
-use ezkl_lib::circuit::base::*;
+use ezkl_lib::circuit::*;
 use ezkl_lib::commands::TranscriptType;
 use ezkl_lib::execute::create_proof_circuit_kzg;
 use ezkl_lib::pfsys::create_keys;
@@ -63,7 +63,8 @@ impl Circuit<Fr> for MyCircuit {
                             padding: (0, 0),
                             stride: (1, 1),
                             kernel_shape: (2, 2),
-                        },
+                        }
+                        .into(),
                     )
                     .unwrap();
                 Ok(())
